hw_04/dags/app/requirements.txt
colorlog
PyYAML
flake8
requests